pieni korjaus 1

+++ versio1.java
@@ -12,7 +12,7 @@ class Versio1 {
 		int kolmikoita = 0;
 		List<int[]> kolmikot = new ArrayList((int)3e6);
 		while (System.currentTimeMillis()-t1 < m*60*1e3
-					&& 2*a*a <= Double.MAX_VALUE )
+					&& 2*a*a <= (int)Double.MAX_VALUE )
 		{
 			for (int b = 1; b <= a; b++) {
 				int s = a*a+b*b;
